reconfiguring auth

+++ app/http/controllers/PCOwnerController.py
@@ -32,6 +32,7 @@ class PCOwnerController(Controller):
         misc_id = self.request.input("misc_id")
         secondmisc_id = self.request.input("secondmisc_id")
         thirdmisc_id = self.request.input("thirdmisc_id")
+        user = self.request.user()
         pc = PC.create({
             "pc_name": pc_name,
             "case_id": case_id,
@@ -46,7 +47,8 @@ class PCOwnerController(Controller):
             "thirdstorage_id":thirdstorage_id,
             "misc_id": misc_id,
             "secondmisc_id": secondmisc_id,
-            "thirdmisc_id": thirdmisc_id
+            "thirdmisc_id": thirdmisc_id,
+            "user_id": user["id"]
         })
         return pc
 
@@ -55,6 +57,7 @@ class PCOwnerController(Controller):
         ex. Model.find('id')
             Get().route("/show", PCOwnerController)
         """
+        user = self.request.user()
         id = self.request.param("id")
         return PC.joins("cpu", "case", "motherboard", "ram", "gpu", "cooler", "psu", "storage", "secondstorage", "thirdstorage", "misc", "secondmisc", "thirdmisc").where("id", id).all()
 
@@ -63,6 +66,7 @@ class PCOwnerController(Controller):
         ex. Post target to update new Model
             Post().route("/update", PCOwnerController)
         """
+        user = self.request.user()
         id = self.request.param("id")
         pc_name = self.request.input("pc_name")
         case_id = self.request.input("case_id")
@@ -78,7 +82,7 @@ class PCOwnerController(Controller):
         misc_id = self.request.input("misc_id")
         secondmisc_id = self.request.input("secondmisc_id")
         thirdmisc_id = self.request.input("thirdmisc_id")
-        PC.where("id", id).update({
+        PC.where("id", id).where("user_id", user["id"]).update({
             "pc_name": pc_name,
             "case_id": case_id,
             "motherboard_id": motherboard_id,
@@ -94,13 +98,14 @@ class PCOwnerController(Controller):
             "secondmisc_id": secondmisc_id,
             "thirdmisc_id": thirdmisc_id
         })
-        return PC.where("id", id).get()
+        return PC.where("id", id).where("user_id", user["id"]).get()
 
     def destroy(self):
         """Delete an existing resource listing
         ex. Delete().route("/destroy", PCOwnerController)
         """
+        user = self.request.user()
         id = self.request.param("id")
-        pc = PC.where("id", id)
-        PC.where("id", id).delete()
+        pc = PC.where("id", id).where("user_id", user["id"])
+        PC.where("id", id).where("user_id", user["id"]).delete()
         return pc
